code/requirements.txt
celery==4.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-cors-headers==3.1.0
django-celery-beat==2.0.0
django-crispy-forms==1.7.2
django-extensions==2.2.9
django-filter==2.2.0
django-formtools==2.1
django-import-export==1.2.0
django-jsonfield==1.3.1
django-mysql==3.8.1
django-redis==4.11.0
django-redis-cache==2.0.0
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.1
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
drf-yasg==1.16.1
future==0.17.1
hiredis==1.0.0
kombu==4.6.6
lxml==4.4.2
MarkupSafe==1.1.1
mysqlclient==1.4.4
openpyxl==2.6.3
Pillow==6.1.0
psutil==5.7.0
pycryptodome==3.9.4
PyJWT==1.7.1
PyMySQL==0.9.3
redis==3.3.8
requests==2.22.0
six==1.12.0
uWSGI==2.0.18
xlrd==1.2.0
xlwt==1.3.0
